rename PremakeDeps to PremakeMultiDeps and add single configuration PremakeDeps
