+++ conan/internal/api/install/generators.py
@@ -29,6 +29,7 @@ _generators = {"CMakeToolchain": "conan.tools.cmake",
                "XcodeDeps": "conan.tools.apple",
                "XcodeToolchain": "conan.tools.apple",
                "PremakeDeps": "conan.tools.premake",
+               "PremakeMultiDeps": "conan.tools.premake",
                "MakeDeps": "conan.tools.gnu",
                "SConsDeps": "conan.tools.scons",
                "QbsDeps": "conan.tools.qbs",
