+++ conan/tools/premake/__init__.py
@@ -1,2 +1,3 @@
 from conan.tools.premake.premake import Premake
 from conan.tools.premake.premakedeps import PremakeDeps
+from conan.tools.premake.premakemultideps import PremakeMultiDeps
